refactor: move ggshield.secret to ggshield.verticals.secret
