+++ ggshield/cmd/secret/scan/ci.py
@@ -14,7 +14,7 @@ from ggshield.core.errors import handle_exception
 from ggshield.core.git_hooks.ci import collect_commit_range_from_ci_env
 from ggshield.core.git_shell import check_git_dir
 from ggshield.scan import ScanContext, ScanMode
-from ggshield.secret.repo import scan_commit_range
+from ggshield.verticals.secret.repo import scan_commit_range
 
 
 @click.command()
