+++ ggshield/cmd/secret/scan/prepush.py
@@ -13,8 +13,8 @@ from ggshield.core.git_hooks.prepush import BYPASS_MESSAGE, collect_commits_refs
 from ggshield.core.git_shell import check_git_dir, get_list_commit_SHA
 from ggshield.core.utils import EMPTY_SHA, EMPTY_TREE
 from ggshield.scan import ScanContext, ScanMode
-from ggshield.secret.output.messages import remediation_message
-from ggshield.secret.repo import scan_commit_range
+from ggshield.verticals.secret.output.messages import remediation_message
+from ggshield.verticals.secret.repo import scan_commit_range
 
 
 logger = logging.getLogger(__name__)
