+++ ggshield/verticals/secret/docker.py
