+++ ggshield/verticals/secret/output/__init__.py
