+++ ggshield/verticals/secret/output/messages.py
