+++ ggshield/verticals/secret/output/schemas.py
