+++ ggshield/verticals/secret/output/secret_gitlab_webui_output_handler.py
