+++ ggshield/verticals/secret/output/secret_json_output_handler.py
