+++ ggshield/verticals/secret/output/secret_text_output_handler.py
