+++ ggshield/verticals/secret/rich_secret_scanner_ui.py
