+++ ggshield/verticals/secret/secret_scan_collection.py
