+++ ggshield/verticals/secret/secret_scanner.py
