+++ tests/unit/cmd/iac/test_scan_diff.py
@@ -4,7 +4,7 @@ from click.testing import CliRunner
 
 from ggshield.cmd.main import cli
 from ggshield.core.errors import ExitCode
-from ggshield.secret.repo import cd
+from ggshield.verticals.secret.repo import cd
 from tests.conftest import _IAC_SINGLE_VULNERABILITY
 from tests.repository import Repository
 from tests.unit.conftest import my_vcr
