+++ tests/unit/cmd/iac/test_scan_prepush.py
@@ -5,7 +5,7 @@ import pytest
 from click.testing import CliRunner
 
 from ggshield.cmd.main import cli
-from ggshield.secret.repo import cd
+from ggshield.verticals.secret.repo import cd
 from tests.conftest import _IAC_NO_VULNERABILITIES, _IAC_SINGLE_VULNERABILITY
 from tests.repository import Repository
 from tests.unit.conftest import my_vcr
