+++ tests/unit/cmd/sca/test_diff.py
@@ -2,7 +2,7 @@ import pytest
 
 from ggshield.cmd.main import cli
 from ggshield.core.errors import ExitCode
-from ggshield.secret.repo import cd
+from ggshield.verticals.secret.repo import cd
 from tests.unit.conftest import my_vcr
