+++ tests/unit/cmd/sca/test_precommit.py
@@ -12,7 +12,7 @@ from ggshield.sca.sca_scan_models import (
     SCAVulnerability,
     SCAVulnerablePackageVersion,
 )
-from ggshield.secret.repo import cd
+from ggshield.verticals.secret.repo import cd
 from tests.repository import Repository
 from tests.unit.conftest import my_vcr
