+++ tests/unit/cmd/sca/test_scan_prereceive.py
@@ -6,7 +6,7 @@ from click.testing import CliRunner
 
 from ggshield.cmd.main import cli
 from ggshield.core.errors import ExitCode
-from ggshield.secret.repo import cd
+from ggshield.verticals.secret.repo import cd
 from tests.repository import create_pre_receive_repo
 from tests.unit.conftest import my_vcr
