+++ tests/unit/cmd/test_ignore.py
@@ -8,7 +8,7 @@ from ggshield.core.cache import Cache
 from ggshield.core.config import Config
 from ggshield.core.types import IgnoredMatch
 from ggshield.scan import Commit, ScanContext, ScanMode
-from ggshield.secret import SecretScanner
+from ggshield.verticals.secret import SecretScanner
 from tests.unit.conftest import _MULTIPLE_SECRETS_PATCH, my_vcr
