+++ tests/unit/core/test_utils.py
@@ -22,8 +22,8 @@ from ggshield.core.utils import (
 )
 from ggshield.scan import Commit, Files, ScanContext, ScanMode, StringScannable
 from ggshield.scan.scan_context import parse_os_release
-from ggshield.secret import SecretScanner
-from ggshield.secret.repo import cd
+from ggshield.verticals.secret import SecretScanner
+from ggshield.verticals.secret.repo import cd
 from tests.unit.conftest import (
     _PATCH_WITH_NONEWLINE_BEFORE_SECRET,
     _SECRET_RAW_FILE,
