+++ tests/unit/verticals/secret/output/snapshots/snap_test_text_output.py
