+++ tests/unit/verticals/secret/output/test_gitlab_webui_output.py
@@ -4,8 +4,8 @@ import pytest
 from pygitguardian.models import Match, PolicyBreak
 
 from ggshield.scan import StringScannable
-from ggshield.secret import Result, Results, SecretScanCollection
-from ggshield.secret.output.secret_gitlab_webui_output_handler import (
+from ggshield.verticals.secret import Result, Results, SecretScanCollection
+from ggshield.verticals.secret.output.secret_gitlab_webui_output_handler import (
     SecretGitLabWebUIOutputHandler,
     format_policy_break,
 )
