+++ tests/unit/verticals/secret/snapshots/snap_test_secret_scanner.py
