+++ tests/unit/verticals/secret/test_scan_docset.py
